apps/api/requirements.txt

fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
pydantic==2.7.1
pydantic-settings==2.2.1
sqlalchemy==2.0.30
geoalchemy2==0.14.6
asyncpg==0.29.0
psycopg2-binary==2.9.9
redis==5.0.4
python-dotenv==1.0.1
rasterio==1.3.10
shapely==2.0.4
numpy==1.26.4
Pillow==10.3.0
pyproj==3.6.1
aiofiles==23.2.1
transformers==4.41.0
accelerate==0.30.0
# CPU-only torch — install from PyTorch's own index, avoids the 779MB CUDA build
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.3.0+cpu
torchvision==0.18.0+cpu
timm==0.9.16